Add changelog script
This little script will return a list of fixed bugs and new features since
the last tag. The bug and release qualifiers have been made quite general,
so they need to be customized to the format used in the repository.

Add it to PATH to be able to 'git changelog' in the terminal